updated user specific routes to check for session data

+++ flask_app/flask06.py
@@ -29,9 +29,9 @@ with app.app_context():
 @app.route('/')
 @app.route('/index')
 def index():
-    a_user =  db.session.query(User).filter_by(email="varunu311@gmail.com").one()
-    print(a_user)
-    return render_template('index.html', user = a_user)
+    if session.get('user'):
+        return render_template('index.html', user = session['user'])
+    return render_template('index.html')
 
 @app.route('/notes')
 def get_notes():
@@ -44,53 +44,57 @@ def get_notes():
 
 @app.route('/notes/new', methods=['GET','POST'])
 def new_note():
+    if session.get('user'):
+        if request.method == 'POST':
+            title = request.form['title']
+            text = request.form['noteText']
 
-    if request.method == 'POST':
-        title = request.form['title']
-        text = request.form['noteText']
+            from datetime import date
+            today = date.today()
 
-        from datetime import date
-        today = date.today()
-
-        today = today.strftime("%m-%d-%Y")
-        new_record = Note(title, text, today)
-        db.session.add(new_record)
-        db.session.commit()
-        return redirect(url_for('get_notes'))
+            today = today.strftime("%m-%d-%Y")
+            new_record = Note(title, text, today)
+            db.session.add(new_record)
+            db.session.commit()
+            return redirect(url_for('get_notes'))
+        else:
+            return render_template('new.html', user=session['user'])
     else:
-        a_user = db.session.query(User).filter_by(email='varunu311@gmail.com').one()
-        return render_template('new.html', user=a_user)
+        return redirect(url_for('login'))
 
 @app.route('/notes/edit/<note_id>', methods=['GET','POST'])
 def update_note(note_id):
-    if request.method == 'POST':
+    if session.get('user'):
+        if request.method == 'POST':
+            
+            title = request.form['title']
+            
+            text = request.form['noteText']
+            note = db.session.query(Note).filter_by(id=note_id).one()
         
-        title = request.form['title']
-        
-        text = request.form['noteText']
-        note = db.session.query(Note).filter_by(id=note_id).one()
-       
-        note.title = title
-        note.text = text
+            note.title = title
+            note.text = text
 
-        db.session.add(note)
-        db.session.commit()
+            db.session.add(note)
+            db.session.commit()
 
-        return redirect(url_for('get_notes'))
+            return redirect(url_for('get_notes'))
+        else:            
+            my_note = db.session.query(Note).filter_by(id=note_id).one()
+            return render_template('new.html',note=my_note,user=session['user'])
     else:
-        a_user = db.session.query(User).filter_by(email='mogli@uncc.edu').one()
-        
-        my_note = db.session.query(Note).filter_by(id=note_id).one()
-        
-        return render_template('new.html',note=my_note,user=a_user)
+        return redirect(url_for('login'))
+
 
 @app.route('/notes/delete/<note_id>', methods = ['POST'])
 def delete_note(note_id):
-    my_note = db.session.query(Note).filter_by(id=note_id).one()
-    db.session.delete(my_note)
-    db.session.commit()
-
-    return redirect(url_for('get_notes'))
+    if session.get('user'):
+        my_note = db.session.query(Note).filter_by(id=note_id).one()
+        db.session.delete(my_note)
+        db.session.commit()
+        return redirect(url_for('get_notes'))
+    else:
+        return redirect(url_for('login'))
 
 @app.route('/notes/<note_id>')
 def get_note(note_id):
